Fix a bf16 bug in the chunk kernel, add solve-rate figure, show raw seeds
The gate maths is kept in fp32 on purpose, but g_C multiplies the carried
state -- leaving it fp32 promoted S on the first chunk and the *second* chunk
then died with 'expected BFloat16 but found Float'. Invisible to the whole
suite, which runs in float64, and invisible at T <= one chunk. Now covered by
an end-to-end low-precision test across fp32/bf16/fp16, and the benchmark can
finally time our readable implementation: 82x slower than the fused kernel.

Figures: MQAR accuracy is bimodal (a seed solves or sits at chance), so the
mean-with-min/max-band panel drew a line through territory no run occupied.
Individual seeds are now plotted as dots, and solve_rate_mqar reports the
statistic the outcome actually supports.

+++ kda/chunk.py
@@ -209,7 +209,10 @@ def chunk_linear_attn(
     # gc[..., r, :] = sum_{j<=r} log a_j   (inclusive), so exp(gc) = g_r.
     gc = la.cumsum(dim=-2)
     g = gc.exp()                       # G, all entries in (0, 1]
-    g_last = g[..., -1:, :]            # g_C
+    # g_C, cast to the compute dtype. The gate maths is deliberately kept wide, but this
+    # value multiplies the carried state: leaving it fp32 promotes S on the first chunk
+    # and the next chunk's q_t @ S then dies with "expected BFloat16 but found Float".
+    g_last = g[..., -1:, :].to(q.dtype)
 
     q_t = (q * g.to(q.dtype))          # Q~ = G * Q
     k_t = (k_delta * g.to(q.dtype))    # K~ = G * K   (delta path only)
